+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed `fibonacci` functions.
 - Fixed `#[criterion]` benchmarks ignoring the command-line options.
 - Fixed incorrect scaling of the violin plots.
+- Don't print the recommended sample count if it's the same as the configured
+  sample count.
 
 ## [0.3.0] - 2019-08-25
 ### Added